Worked on Java Doc and started on decode implementation. We need to wrap up decode tomorrow and start on the encode implementations. We also have to make sure we are writing tests along with these implementations.

+++ src/notifi/serialization/NoTiFiLocationAddition.java
@@ -12,7 +12,7 @@ public class NoTiFiLocationAddition extends NoTiFiMessage{
     /**
      * The Operation Code for the NoTiFi Location Addition
      */
-    static final int LOCATION_ADDITION_CODE = 1;
+    static final short LOCATION_ADDITION_CODE = 1;
 
 
     /**
@@ -98,6 +98,7 @@ public class NoTiFiLocationAddition extends NoTiFiMessage{
     //TODO Encode implementation
     @Override
     public byte[] encode() {
+
         return new byte[0];
     }
